Restructured code into one-file-per-class, regenerated .gemspec
